bee-hive/pyproject.toml
[tool.poetry]
name = "bee_hive"
version = "0.1.0"
description = "A multi-agent platform with the vision to facilitate deploy and run Bee agents."
authors = ["IBM"]
license = "Apache 2.0"
readme = "README.md"
packages = [
    { include = "bee_hive", from = "." },
    { include = "cli", from = "." }
]

[tool.poetry.dependencies]
python = ">= 3.11, < 3.13"
pyyaml = "^6.0.2"
openai = "^1.56.2"
python-dotenv = "^1.0.1"
jsonschema = "^4.23.0"
docopt-ng = "^0.9.0"

[tool.poetry.group.dev.dependencies]

[tool.poetry.group.test.dependencies]
pytest-mock = "^3.14.0"
pytest = "^8.3.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
addopts = "-v -s"

[tool.poetry.scripts]
"beeAI" = "cli.beeAI:run_cli"
